requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file='.\requirements\requirements.txt' '.\requirements\requirements.in'
#
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
boto3==1.37.37
    # via -r .\requirements\requirements.in
botocore==1.37.37
    # via
    #   boto3
    #   s3transfer
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via
    #   prisma
    #   rich-toolkit
    #   typer
    #   uvicorn
colorama==0.4.6
    # via
    #   click
    #   uvicorn
dnspython==2.7.0
    # via email-validator
email-validator==2.2.0
    # via fastapi
exceptiongroup==1.2.2
    # via anyio
fastapi[standard]==0.115.12
    # via -r .\requirements\requirements.in
fastapi-cli[standard]==0.0.7
    # via fastapi
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.8
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via
    #   fastapi
    #   prisma
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
jinja2==3.1.6
    # via
    #   fastapi
    #   prisma
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
nodeenv==1.9.1
    # via prisma
prisma==0.15.0
    # via -r .\requirements\requirements.in
pydantic==2.11.3
    # via
    #   -r .\requirements\requirements.in
    #   fastapi
    #   prisma
pydantic-core==2.33.1
    # via pydantic
pygments==2.19.1
    # via rich
python-dateutil==2.9.0.post0
    # via botocore
python-dotenv==1.1.0
    # via
    #   -r .\requirements\requirements.in
    #   prisma
    #   uvicorn
python-multipart==0.0.20
    # via fastapi
pyyaml==6.0.2
    # via uvicorn
requests==2.32.3
    # via -r .\requirements\requirements.in
rich==14.0.0
    # via
    #   rich-toolkit
    #   typer
rich-toolkit==0.14.1
    # via fastapi-cli
s3transfer==0.11.5
    # via boto3
shellingham==1.5.4
    # via typer
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
starlette==0.46.2
    # via fastapi
strenum==0.4.15
    # via prisma
tomlkit==0.13.2
    # via prisma
typer==0.15.2
    # via fastapi-cli
typing-extensions==4.13.2
    # via
    #   anyio
    #   fastapi
    #   prisma
    #   pydantic
    #   pydantic-core
    #   rich
    #   rich-toolkit
    #   typer
    #   typing-inspection
    #   uvicorn
typing-inspection==0.4.0
    # via pydantic
urllib3==2.4.0
    # via
    #   botocore
    #   requests
uvicorn[standard]==0.34.2
    # via
    #   -r .\requirements\requirements.in
    #   fastapi
    #   fastapi-cli
watchfiles==1.0.5
    # via uvicorn
websockets==15.0.1
    # via uvicorn
